+++ Source/Public/Entra/Get-IRTServicePrincipalSignInLog.ps1
@@ -87,7 +87,7 @@ function Get-IRTServicePrincipalSignInLog {
         [boolean] $Beta = $true,
         [boolean] $Excel = $true,
         [boolean] $IpInfo = [bool]$Global:IRT_Config.IpInfoAvailable,
-        [boolean] $Open = $true,
+        [boolean] $Open = [bool]$Global:IRT_Config.OpenSpreadsheets,
         [boolean] $Xml = $Global:IRT_Config.ExportXml
     )