+++ Source/Public/Entra/Show-IRTServicePrincipalSignIn.ps1
@@ -47,7 +47,7 @@ function Show-IRTServicePrincipalSignIn {
         [string]  $Font = $Global:IRT_Config.ExcelFont,
 
         [boolean] $IpInfo = [bool]$Global:IRT_Config.IpInfoAvailable,
-        [boolean] $Open = $true
+        [boolean] $Open = [bool]$Global:IRT_Config.OpenSpreadsheets
     )
 
     begin {