+++ Source/Public/Role/Get-IRTAdminRole.ps1
@@ -60,7 +60,7 @@ function Get-IRTAdminRole {
         [string[]] $Highlight,
         [string] $TableStyle = $Global:IRT_Config.ExcelTableStyle,
         [string] $Font = $Global:IRT_Config.ExcelFont,
-        [boolean] $Open = $true
+        [boolean] $Open = [bool]$Global:IRT_Config.OpenSpreadsheets
     )
 
     begin {