+++ Source/Public/ServicePrincipal/Get-IRTUserServicePrincipal.ps1
@@ -56,7 +56,7 @@ function Get-IRTUserServicePrincipal { # FIXME rename to Get-IRTUserAppConsent
         [string] $TableStyle = $Global:IRT_Config.ExcelTableStyle,
         [string] $Font = $Global:IRT_Config.ExcelFont,
         [boolean] $Xml = $Global:IRT_Config.ExportXml,
-        [boolean] $Open = $true,
+        [boolean] $Open = [bool]$Global:IRT_Config.OpenSpreadsheets,
         [switch] $Cached
     )